+++ CHANGES.md
@@ -10,7 +10,6 @@ Cloned on GitHub from the Google Code version
 **Note**: the command line options have changed!  
 **Be careful with existing shell scripts after updating from v.0.x!** 
 
-### Done
   - Support for Python 2.6+ and 3.2+
   - `-x`, `--execute` option was removed. It is now on by default.
     Use `--dry-run` to force simulation mode.
@@ -19,9 +18,8 @@ Cloned on GitHub from the Google Code version
   - `-i`, `--ignore` option was added to allow exclusion patterns.
   - Verbositiy was changed. By default only changed files will be printed.
   - PEP 8
-  - `-m` option now accepts a comma separated list of extensions.
   - Multiple target folders may be passed
-
+  - `-m` option now accepts a comma separated list of patterns.
 
 ## 0.1.7 and before
 Hosted on Google Code: https://code.google.com/p/tabfix/